Make use of libosmcoore cpu_sched_vty
osmo-e1d automatically uses SCHED_RR already, but for consistency it
actually makes sense to add this, so people can configure the scheduler
priority and/or the cpu affinity this way.

The existing call to SCHED_RR is left in place for backwards
compatibility.  The new VTY commands can just override that, if needed.

Change-Id: I1b8021e6dc864af6e301fced427e24c8bb21ca2f