support for e1-tracer hardware
Recent work on the e1-tracer firmware is introducing a set of USB
descriptors (as configuration 2) that are mostly compatible to how
osmo-e1d talks to icE1usb.

The main difference is that there's only one ISO IN endpoint per USB
interface, and no ISO OUT or ISU FB endpoints.  So we introduce some
minor adjustments here to accommodate that.

Related: OS#5733
Closes: OS#5734
Change-Id: I855e18c0f229bd473123f96303e60ab2de90677f