usb: Deal with truncated ISO IN transfers
It seems that in some circumstances, an ISO IN transfer can be
truncated by the bus / host.  In such situation we'd currently pass
a non-modulo-32 length to the mux_demux (deframer) code, and it ASSERTs
on that.  Let's try to handle this more gracefully by substituting
random garbage and letting higher layers deal with massive bit errors.

Related: OS#5490
Change-Id: Ic453325b93b0e12727625a1495a948d96df4b542