mux from not-started raw TS: fill with 0xFF
When the channelized mux reads from a raw TS, but that raw TS is
not in tx_started state yet, the read function returns "fake" data.
However, it was actually returning uninitialized memory content,
rather than 0xFF filler used everywhere else, thereby transmitting
uninit-memory garbage to whatever is connected to the E1 line.
Change it to fill with 0xFF, same as the filler used in other cases
such as inactive timeslots.

Change-Id: I42849a6d19b020bab789853c3b60af6a1c09f92f