Add frame_base to mux/demux functions
This will be required for channels that need to be aligned with framing,
like CAS signaling.

The frame_base is also sent towards USB, to keep the icE1usb in sync
with the host. This is currently ignored. If there is no buffer
overflow / underrun on the USB, the icE1usb should be in sync anyway.

Change-Id: I3a303a9942eb746c90a13cdf9d2930db69c29906